git-svn-id: http://ironman.agnux.com/repositorios/ErpAgnuxV1@1566 ce483f98-b20a-4c78-8f99-c15376fc4320

+++ erp-quimicos/src/main/java/com/agnux/kemikal/controllers/ComOcDiasEntController.java
@@ -257,7 +257,9 @@ public class ComOcDiasEntController {
         FileCopyUtils.copy(bis, response.getOutputStream());  	
         response.flushBuffer();
         
-        FileHelper.delete(fileout);
+        if(file.exists()){
+            FileHelper.delete(fileout);
+        }
         
         return null;
     }
