+++ erp-quimicos/src/main/java/com/agnux/kemikal/controllers/CxpRepIepsPagadoController.java
@@ -537,7 +537,9 @@ public class CxpRepIepsPagadoController {
         FileCopyUtils.copy(bis, response.getOutputStream());  	
         response.flushBuffer();
         
-        FileHelper.delete(fileout);
+        if(file.exists()){
+            FileHelper.delete(fileout);
+        }
         
         return null;
     }
