+++ erp-quimicos/src/main/java/com/agnux/kemikal/controllers/PocPedidosController.java
@@ -1102,7 +1102,9 @@ public class PocPedidosController {
         FileCopyUtils.copy(bis, response.getOutputStream());          
         response.flushBuffer();
         
-        FileHelper.delete(fileout);
+        if(file.exists()){
+            FileHelper.delete(fileout);
+        }
         
         return null;
